refactor(killers): create killers struct and rename a few files
Bench: 21585
