+++ engine/src/search/info.rs
