+++ engine/src/search/params.rs
@@ -22,4 +22,3 @@ impl Default for SearchParams {
         }
     }
 }
-
