+++ engine/src/search/tt.rs
@@ -124,4 +124,3 @@ impl Default for TranspositionTableEntry {
         }
     }
 }
-
